Returns HttpError object for HTTP status != 200
If the HTTP status returned is not 200, return a new HttpError object.  This object has two fields:  StatusCode that shows the HTTP status; and error which is the error object returned prior to this change.  This allows caller to check the HTTP status code without parsing the actual error text.